api/requirements.txt
# Core Flask and API framework
Flask==3.0.0
flask-openapi3==3.1.1
flask-cors==4.0.0

# Data validation and serialization
pydantic==2.5.2
pydantic[email]==2.5.2

# Authentication and security
PyJWT==2.8.0
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
cryptography>=41.0.0

# Database and caching
pymongo==4.6.1
upstash-redis==0.15.0
redis==5.0.1

# Message queue
pika==1.3.2

# OpenTelemetry observability (using latest stable versions)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-flask>=0.41b0
opentelemetry-instrumentation-pymongo>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-exporter-otlp>=1.20.0

# HTTP client
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
jsonpath-ng==1.6.1
psutil==5.9.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
responses==0.24.1
testcontainers==3.7.1

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Production server
gunicorn==21.2.0
uvicorn==0.24.0